+++ packages/cli/src/controllers/api-keys.controller.ts
@@ -1,11 +1,11 @@
 import { CreateApiKeyRequestDto, UpdateApiKeyRequestDto } from '@n8n/api-types';
 import { Body, Delete, Get, Param, Patch, Post, RestController } from '@n8n/decorators';
+import { getApiKeyScopesForRole } from '@n8n/permissions';
 import type { RequestHandler } from 'express';
 
 import { BadRequestError } from '@/errors/response-errors/bad-request.error';
 import { EventService } from '@/events/event.service';
 import { isApiEnabled } from '@/public-api';
-import { getApiKeyScopesForRole } from '@/public-api/permissions.ee';
 import { AuthenticatedRequest } from '@/requests';
 import { PublicApiKeyService } from '@/services/public-api-key.service';
