+++ packages/cli/src/services/public-api-key.service.ts
@@ -4,6 +4,7 @@ import type { User } from '@n8n/db';
 import { ApiKey, ApiKeyRepository, UserRepository } from '@n8n/db';
 import { Service } from '@n8n/di';
 import type { GlobalRole, ApiKeyScope } from '@n8n/permissions';
+import { getApiKeyScopesForRole, getOwnerOnlyApiKeyScopes } from '@n8n/permissions';
 // eslint-disable-next-line n8n-local-rules/misplaced-n8n-typeorm-import
 import type { EntityManager } from '@n8n/typeorm';
 import type { NextFunction, Request, Response } from 'express';
@@ -11,7 +12,6 @@ import { TokenExpiredError } from 'jsonwebtoken';
 import type { OpenAPIV3 } from 'openapi-types';
 
 import { EventService } from '@/events/event.service';
-import { getApiKeyScopesForRole, getOwnerOnlyApiKeyScopes } from '@/public-api/permissions.ee';
 import type { AuthenticatedRequest } from '@/requests';
 
 import { JwtService } from './jwt.service';
