+++ packages/cli/test/integration/api-keys.api.test.ts
@@ -3,11 +3,14 @@ import { GlobalConfig } from '@n8n/config';
 import type { User } from '@n8n/db';
 import { ApiKeyRepository } from '@n8n/db';
 import { Container } from '@n8n/di';
-import type { ApiKeyScope } from '@n8n/permissions';
+import {
+	getApiKeyScopesForRole,
+	getOwnerOnlyApiKeyScopes,
+	type ApiKeyScope,
+} from '@n8n/permissions';
 import { mock } from 'jest-mock-extended';
 
 import type { License } from '@/license';
-import { getApiKeyScopesForRole, getOwnerOnlyApiKeyScopes } from '@/public-api/permissions.ee';
 import { PublicApiKeyService } from '@/services/public-api-key.service';
 import { mockInstance } from '@test/mocking';
